feat: native inplementation
